pyproject.toml
[project]
name = "aumilas-ae-3-5"
version = "0.1.0"
description = ""
authors = [
    {name = "milaura22@gmail.com",email = "milaura22@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "streamlit (>=1.44.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "openai (>=1.70.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "langchain-openai (>=0.3.12,<0.4.0)",
    "langchain-community (>=0.3.21,<0.4.0)"
]

[tool.poetry]
packages = [{include = "aumilas_ae"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
